export and build fixes
- moved imported models from Blocks.blend to imported (because .godot/imported is not exported correctly)
- fixed / updated builds libs for windows
  - added pacman msys2 support for download windows dependencies (ngspice and qemu)
- updated OS image build script to Debian Trixie
  - droped initrd mode
  - added multiuser support
- added download Linux sources to ComputerSimulator/OS/Makefile
  - now `just` command should automatically build all project dependencies on clean repo
- added support for export Linux ngspice and qemu binaries
- used dlopen in GdSpice also for Windows build
- export copyright info for binary dependencies

+++ Justfile
@@ -5,22 +5,37 @@ run: build
 	godot project.godot
 
 build: init-submodules build-gdspice build-qemu-images build-addons-godot-xterm build-manual build-addons-gdcef
+	mkdir -p imported
 
 build-windows-libs:
-	@ WINDOWS=true just build
+	@ WINDOWS=true just build-gdspice build-addons-godot-xterm build-addons-gdcef
 
 init-submodules:
 	#!/bin/sh -e
 	echo '{{BOLD + YELLOW}} Init git submodules {{NORMAL}}'
 	git submodule update --init
 
-build-gdspice: build-addons-godot-xterm 
+build-godot-cpp: init-submodules
+	#!/bin/sh -e
+	echo '{{BOLD + YELLOW}} Build godot-cpp {{NORMAL}}'
+	cd addons/3rdparty/_godot-cpp/
+	git submodule update --init .
+	if [ "$WINDOWS" = "true" ]; then
+		[ -f bin/libgodot-cpp.windows.template_debug.x86_64.a ] || scons platform=windows arch=x86_64
+		[ -f bin/libgodot-cpp.windows.template_release.x86_64.a ] || ln -sr bin/libgodot-cpp.windows.template_debug.x86_64.a bin/libgodot-cpp.windows.template_release.x86_64.a
+	else
+		[ -f bin/libgodot-cpp.linux.template_debug.x86_64.a ] || scons
+		[ -f bin/libgodot-cpp.linux.template_release.x86_64.a ] || ln -sr bin/libgodot-cpp.linux.template_debug.x86_64.a bin/libgodot-cpp.linux.template_release.x86_64.a
+	fi
+
+build-gdspice: build-godot-cpp
 	#!/bin/sh -e
 	echo '{{BOLD + YELLOW}} Build gdSpice {{NORMAL}}'
 	cd ElectronicsSimulator/GdSpice/
-	scons
 	if [ "$WINDOWS" = "true" ]; then
 		scons platform=windows arch=x86_64
+	else
+		scons
 	fi
 
 build-qemu-images:
@@ -29,14 +44,15 @@ build-qemu-images:
 	cd ComputerSimulator/OS/
 	make
 
-build-addons-godot-xterm: init-submodules
+build-addons-godot-xterm: build-godot-cpp
 	#!/bin/sh -e
 	echo '{{BOLD + YELLOW}} Build GodotXterm {{NORMAL}}'
 	cd addons/3rdparty/godot-xterm/
-	just
 	if [ "$WINDOWS" = "true" ]; then
 		cd addons/godot_xterm/native/
-		scons platform=windows arch=x86_64
+		scons platform=windows arch=x86_64 build_library=False no_pty=True
+	else
+		SCONSFLAGS="build_library=False" just
 	fi
 
 build-manual:
@@ -45,31 +61,45 @@ build-manual:
 	cd Manual/
 	scons
 
-build-addons-gdcef: init-submodules
+build-addons-gdcef: build-godot-cpp
 	#!/bin/sh -e
 	echo '{{BOLD + YELLOW}} Build gdCef {{NORMAL}}'
 	cd addons/3rdparty/gdcef/
 	
-	(
-		cd addons/gdcef/
-		while read p i; do
-			pip3 show $p > /dev/null
-		done < requirements.txt
+	if [ "$WINDOWS" = "true" ]; then
+		(
+			cd build/
+			if [ -f Windows/GDCEF_VERSION.txt ]; then
+				exit
+			fi
+			wget https://github.com/Lecrapouille/gdcef/releases/download/v0.17.0-godot4/gdCEF-0.17.0_Godot-4.3_Windows_X64.tar.gz
+			tar -xzf gdCEF-0.17.0_Godot-4.3_Windows_X64.tar.gz
+			rm -fr build/Windows
+			mv cef_artifacts/ Windows
+			rm gdCEF-0.17.0_Godot-4.3_Windows_X64.tar.gz
+		)
+	else
+		(
+			cd addons/gdcef/
+			
+			if [ ! -d thirdparty/godot-4.3/cpp/ ]; then
+				mkdir -p thirdparty/godot-4.3/
+				ln -s {{justfile_directory()}}/addons/3rdparty/_godot-cpp/ thirdparty/godot-4.3/cpp
+			fi
+			
+			while read p i; do
+				pip3 show $p > /dev/null
+			done < requirements.txt
+			
+			/usr/bin/env python3 build.py
+		)
 		
-		/usr/bin/env python3 build.py
-	)
-	
-	mkdir -p build/
-	rm -fr build/Linux
-	mv cef_artifacts/ build/Linux
-	
-	(cd  build/; ln -sf Linux _current_platform_)
-	
-	if [ "$WINDOWS" = "true" ]; then (
-		cd build/
-		wget https://github.com/Lecrapouille/gdcef/releases/download/v0.17.0-godot4/gdCEF-0.17.0_Godot-4.3_Windows_X64.tar.gz
-		tar -xzf gdCEF-0.17.0_Godot-4.3_Windows_X64.tar.gz
-		rm -fr build/Windows
-		mv cef_artifacts/ Windows
-		rm gdCEF-0.17.0_Godot-4.3_Windows_X64.tar.gz
-	); fi
+		mkdir -p build/
+		rm -fr build/Linux
+		mv cef_artifacts/ build/Linux
+		
+		(
+			cd  build/
+			ln -sf Linux _current_platform_
+		)
+	fi
